requirements.txt
urllib3
openai
fastapi
uvicorn
requests
python-dotenv
tiktoken
pydantic
pytest
langchain
langdetect
pillow
sentence-transformers
numpy
gitpython
duckdb
beautifulsoup4
speechrecognition
pydub
markdown
joblib
langchain-community
pandas
scikit-learn
httpx
speechrecognition
pydub
openai-whisper
GitPython
transformers
lxml
python-dateutil
faker